+++ crates/webidl-tests/lib.rs
@@ -1,0 +1,1 @@
+// intentionally left blank
