+++ crates/webidl-tests/main.rs
@@ -1,0 +1,10 @@
+#![feature(use_extern_macros)]
+
+extern crate wasm_bindgen_test;
+extern crate wasm_bindgen;
+
+pub mod consts;
+pub mod enums;
+pub mod simple;
+pub mod throws;
+pub mod array;
