+++ crates/webidl/Cargo.toml
@@ -11,9 +11,6 @@ description = """
 Support for parsing WebIDL specific to wasm-bindgen
 """
 
-[dev-dependencies]
-wasm-bindgen-test-project-builder = { path = "../test-project-builder", version = '=0.2.15' }
-
 [dependencies]
 failure = "0.1"
 failure_derive = "0.1"
