--- crates/webidl/tests/all/main.rs
@@ -1,8 +1,0 @@
-extern crate wasm_bindgen_test_project_builder as project_builder;
-use project_builder::project;
-
-mod consts;
-mod enums;
-mod simple;
-mod throws;
-mod array;
